refactor: address review comments for omen components
